% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_2016.R
\docType{data}
\name{calibration_2016}
\alias{calibration_2016}
\title{Dataframe object with solar radiation series for calibration.}
\format{An object of class \code{"data.frame"}.}
\source{
BSRN and CAMSRad service
}
\usage{
data(calibration_2016)
}
\description{
The Dataframe includes time, modeled and measured solar irradiance,
modeled clear sky index and solar elevation (degrees).
}
\references{
Fernandez-Peruchena, C. M. et al (2020).
Site-adaptation of modeled solar radiation data: The SiteAdapt procedure.
Remote Sensing.
}
\keyword{datasets}
