% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_Kasten.r
\name{m_Kasten}
\alias{m_Kasten}
\title{Calculation of relative air mass based on Kasten parametrization}
\usage{
m_Kasten(Sun_elev, z)
}
\arguments{
\item{Sun_elev}{Sun elevation angle (degrees) above horizon}

\item{z}{Site elevation above sea level (m)}
}
\value{
Relative air mass based on Kasten parametrization
}
\description{
Calculation of relative air mass based on Kasten parametrization
}
