% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLIS.r
\name{SOLIS}
\alias{SOLIS}
\title{DNI under clear sky conditions (Solar Energy, 82(8), 758-762)}
\usage{
SOLIS(top, Sun_elev, aod_380, aod_500, w, site_elevation)
}
\arguments{
\item{top}{Solar irradiance at the top of atosphere}

\item{Sun_elev}{Sun elevation angle (degrees) above horizon}

\item{aod_380}{Aerosol optical depth at 380 nm (dimensionless)}

\item{aod_500}{Aerosol optical depth at 500 nm (dimensionless)}

\item{w}{Atmospheric water vapor content (cm)}

\item{site_elevation}{Site elevation above sea level (m)}
}
\value{
Dataframe object including time and site adapted solar irradiance series
}
\description{
DNI under clear sky conditions (Solar Energy, 82(8), 758-762)
}
