% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_2013_2016.R
\docType{data}
\name{target_2013_2016}
\alias{target_2013_2016}
\title{Dataframe object with modeled solar radiation series.}
\format{An object of class \code{"data.frame"}.}
\source{
BSRN
}
\usage{
data(target_2013_2016)
}
\description{
The Dataframe includes time, modeled solar irradiance, modeled
clear sky index and solar elevation (degrees).
}
\references{
Fernandez-Peruchena, C. M. et al (2020).
Site-adaptation of modeled solar radiation data: The SiteAdapt procedure.
Remote Sensing.
}
\keyword{datasets}
