% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOA.r
\name{TOA}
\alias{TOA}
\title{Calculation of the Top of Atmosphere (TOA) solar irradiance on a horizontal plane}
\usage{
TOA(latitude, Sun_elev, Time_Stamp)
}
\arguments{
\item{latitude}{Site latitude (degrees, +N)}

\item{Sun_elev}{Sun elevation angle (degrees) above horizon}

\item{Time_Stamp}{Time series (object of class "POSIXct")}
}
\value{
Top of Atmosphere (TOA) Solar irradiance on a horizontal plane
}
\description{
Calculation of the Top of Atmosphere (TOA) solar irradiance on a horizontal plane
}
