% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_2013_2016.R
\docType{data}
\name{observed_2013_2016}
\alias{observed_2013_2016}
\title{Dataframe object with ground measured solar radiation.}
\format{An object of class \code{"data.frame"}.}
\source{
BSRN
}
\usage{
data(observed_2013_2016)
}
\description{
The Dataframe includes time and solar irradiance measured series.
}
\references{
Fernandez-Peruchena, C. M. et al (2020).
Site-adaptation of modeled solar radiation data: The SiteAdapt procedure.
Remote Sensing.
}
\keyword{datasets}
